Refactor Atomic class package
Move the Atomic class to its new package `jromp.parallel.construct.atomic` for better modularity and clarity. Update `AtomicTests` to import the new package location.
